+++ src/Humanizer/Localisation/Formatters/RussianFormatter.cs
@@ -1,25 +1,23 @@
-﻿namespace Humanizer.Localisation.Formatters
+﻿using Humanizer.Localisation.GrammaticalNumber;
+
+namespace Humanizer.Localisation.Formatters
 {
     internal class RussianFormatter : DefaultFormatter
     {
-        private const string SingularPostfix = "_Singular";
-        private const string PaucalPostfix = "_Paucal";
-
         protected override string GetResourceKey(string resourceKey, int number)
         {
-            var mod100 = number%100;
-            if (mod100/10 != 1)
-            {
-                var mod10 = number%10;
+            var grammaticalNumber = RussianGrammaticalNumberDetector.Detect(number);
+            var suffix = GetSuffix(grammaticalNumber);
+            return resourceKey + suffix;
+        }
 
-                if (mod10 == 1) // 1, 21, 31, 41 ... 91, 101, 121 ..
-                    return resourceKey + SingularPostfix;
-
-                if (mod10 > 1 && mod10 < 5) // 2, 3, 4, 22, 23, 24 ...
-                    return resourceKey + PaucalPostfix;
-            }
-            
-            return resourceKey;
+        private string GetSuffix(RussianGrammaticalNumber grammaticalNumber)
+        {
+            if (grammaticalNumber == RussianGrammaticalNumber.Singular)
+                return "_Singular";
+            if (grammaticalNumber == RussianGrammaticalNumber.Paucal)
+                return "_Paucal";
+            return "";
         }
     }
 }