+++ src/Humanizer/Localisation/GrammaticalNumber/RussianGrammaticalNumber.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+
+namespace Humanizer.Localisation.GrammaticalNumber
+{
+    internal enum RussianGrammaticalNumber
+    {
+        Singular,
+        Paucal,
+        Plural
+    }
+}