+++ src/Humanizer/Localisation/NumberToWords/INumberToWordsConverter.cs
@@ -3,7 +3,7 @@
     /// <summary>
     /// An abstraction to convert number to words
     /// </summary>
-    public interface INumberToWordsConverter
+    internal interface INumberToWordsConverter
     {
         /// <summary>
         /// 3501.ToWords() -> "three thousand five hundred and one"
@@ -11,6 +11,16 @@
         /// <param name="number">Number to be turned to words</param>
         /// <returns></returns>
         string Convert(int number);
+
+        /// <summary>
+        /// for Russian locale
+        /// 1.ToWords(GrammaticalGender.Masculine) -> "один"
+        /// 1.ToWords(GrammaticalGender.Feminine) -> "одна"
+        /// </summary>
+        /// <param name="number">Number to be turned to words</param>
+        /// <param name="gender">The grammatical gender to use for output words</param>
+        /// <returns></returns>
+        string Convert(int number, GrammaticalGender gender);
 
         /// <summary>
         /// 1.ToOrdinalWords() -> "first"